backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
# email-validator required by pydantic EmailStr used in RegisterRequest / UserResponse
email-validator==2.1.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
# python-jose 3.4.0 patches algorithm confusion with OpenSSH ECDSA keys (CVE)
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
# python-multipart 0.0.22 patches arbitrary file write + DoS via malformed boundary (CVE)
python-multipart==0.0.22
celery[redis]==5.4.0
redis==5.0.4
docker==7.0.0
pydantic==2.7.1
pydantic-settings==2.2.1
minio==7.2.7
python-dateutil==2.9.0
httpx==0.27.0
PyYAML==6.0.1
paramiko==3.4.0
pyotp==2.9.0
cryptography==42.0.8
